Agentic_GraphRAG/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
langgraph>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
anthropic>=0.28.0
neo4j>=5.20.0
chromadb>=0.5.0
langfuse>=2.36.0
python-dotenv>=1.0.0
